requirements.txt
ctransformers==0.2.5
# sentence-transformers==2.2.2
pinecone-client
langchain==0.1.9
flask
pypdf
python-dotenv
# -e .